Add UI functionality
-Added functionality for the UI to fill the elements with the data retrieved with the API

+++ app/src/main/java/com/zabava/weatherapp/models/Sys.kt
@@ -6,6 +6,6 @@ data class Sys (
     val type: Int,
     val id: Double,
     val country: String,
-    val sunrise: Double,
-    val sunset: Double,
+    val sunrise: Long,
+    val sunset: Long,
         ):Serializable
